requirements.txt
pytest
Appium-Python-Client
pytest-variables
